refs #47022 Clear TYPO3 configuration cache instead of DCE cache, when saving dce. Clearing of configuration cache also clears dce cache files.
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/dce/trunk@74325 735d13b6-9817-0410-8766-e36946ffe9aa

+++ Classes/Hooks/tx_clearCache.php
@@ -39,8 +39,12 @@ class tx_clearCache {
 	public function clearDceCache($params) {
 		if ($params['cacheCmd'] === 'all' || $params['cacheCmd'] === 'temp_cached') {
 			if (t3lib_utility_VersionNumber::convertVersionNumberToInteger(TYPO3_version) >= 6000000) {
-				unlink($GLOBALS['TYPO3_CONF_VARS']['USER']['dce']['dceLocalconfPath']);
-				unlink($GLOBALS['TYPO3_CONF_VARS']['USER']['dce']['dceExtTablesPath']);
+				if (file_exists($GLOBALS['TYPO3_CONF_VARS']['USER']['dce']['dceLocalconfPath'])) {
+					unlink($GLOBALS['TYPO3_CONF_VARS']['USER']['dce']['dceLocalconfPath']);
+				}
+				if (file_exists($GLOBALS['TYPO3_CONF_VARS']['USER']['dce']['dceExtTablesPath'])) {
+					unlink($GLOBALS['TYPO3_CONF_VARS']['USER']['dce']['dceExtTablesPath']);
+				}
 			} else {
 				t3lib_extMgm::removeCacheFiles('temp_CACHED_dce');
 			}
